requirements.txt
gvm-tools==2.0.0
python-nmap~=0.7.1
python-owasp-zap-v2.4==0.0.19

pyfiglet~=0.8.post1
termcolor~=2.3.0
json2html~=1.3.0
pandas~=2.0.1